+++ Services/OmniSearch.php
@@ -3,22 +3,38 @@
 namespace Leantime\Plugins\OmniSearch\Services;
 
 class OmniSearch {
+  private static $assets = [
+    // source => target
+    __DIR__. '/../assets/js/omniSearch.js' => APP_ROOT . '/public/dist/js/omniSearch.js',
+    __DIR__. '/../assets/css/omniSearch.css' => APP_ROOT . '/public/dist/css/omniSearch.css',
+  ];
 
-  public function install() {
-    symlink(dirname(__DIR__). "/assets/js/omniSearch.js", $_SERVER['DOCUMENT_ROOT'] . "/dist/js/omniSearch.js");
-    symlink(dirname(__DIR__). "/assets/css/omniSearch.css", $_SERVER['DOCUMENT_ROOT'] . "/dist/css/omniSearch.css");
+  /**
+   * Install plugin.
+   *
+   * @return void
+   */
+  public function install(): void
+  {
+    foreach (static::$assets as $source => $target) {
+      if (file_exists($target)) {
+        unlink($target);
+      }
+      symlink($source, $target);
+    }
   }
 
-  public function uninstall() {
-    $omniSelectFiles = [
-        $_SERVER['DOCUMENT_ROOT'] . "/dist/js/omniSearch.js",
-        $_SERVER['DOCUMENT_ROOT'] . "/dist/css/omniSearch.css"
-      ];
-
-      foreach ($omniSelectFiles as $file) {
-        if (file_exists($file)) {
-          unlink($file);
-        }
+  /**
+   * Uninstall plugin.
+   *
+   * @return void
+   */
+  public function uninstall(): void
+  {
+    foreach (static::$assets as $target) {
+      if (file_exists($target)) {
+        unlink($target);
       }
+    }
   }
 }