+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Used `APP_ROOT` for installing assets.
+
 ## [1.2.0] 2024-02-12
 
 ### Fixed